json Word Loader Start
I have implemented the base of the json word loader code. I still need to split up the working class into parts for use through out the project, but the functionality is there. It can load a file that contains a list of file names, load those files to get the word lists, then assemble a "safe" dictonary of keywords and their Synonyms.

The next step is to split up the tmp class functions into parts as some of them need to be called seperatly by other sections of the program. This was just a skeleton class to get the functionality working.

Then I will work on the command processor.